Satisfy mypy
Note that I marked the csv upload function, which is a Flask view
function, as returning `Any`. Even though Flask views are allowed to
return `dict`, the typeshed stub for Flask doesn't include `dict` as an
allowed return type for Flask views. The suggested workaround is to
label such functions as having return type `Any`:
https://github.com/python/mypy/issues/7187#issuecomment-512750396.