Update file type descriptions
I tried to embed hyperlinks into the hint text, but that didn't work.
Curiously, it seems to work on very early versions of Vuetify (according
to an experiment I tried on codepen). If there's a way to do this now, I
would like to, but otherwise it's no big deal.