Add a `session_user()` function
This commit also moves the definition of MULTINET_COOKIE to user.py,
since it's a standalone module that doesn't depend on other modules.
This may not be the best place for this definition, but leaving it in
auth/__init__.py causes a circular import with this module, as does
elevating it to multinet/__init__.py (because that module needs to load
the google oauth blueprint, which in turn needs to make use of the
MULTINET_COOKIE value).