Set up full workspace metadata on creation
This commit features a specific semantics for logged-in vs. logged-out
users creating workspaces: a logged-in user creates a new, non-public
workspace owned by them; a logged-out user simply creates an ownerless,
public workspace.

This is probably the wrong semantics, so we will need to revisit this
later. In particular, we need a way to prevent logged out users from
being able to simply create workspaces endlessly in a kind of anonymous
DoS attack against the database.