Add workspace permisssion user object transform
These functions are temporary solutions to the problem of our permission
storage. Right now, we only store the `sub` value of a user, which means
not all of the information we need is available on the client. The
future solution to this is to store whole user objects in the
`permissions` field of a workspace. However, doing that now would
conflict with future work, so this is the temporary solution.